refactor: move functions out of controller
