+++ apps/omg_watcher_info/lib/omg_watcher_info/api/transaction.ex
@@ -47,8 +47,6 @@ defmodule OMG.WatcherInfo.API.Transaction do
           fee: UtxoSelection.fee_t()
         }
 
-  @type utxos_map_t() :: %{UtxoSelection.currency_t() => UtxoSelection.utxo_list_t()}
-  @type inputs_t() :: {:ok, utxos_map_t()} | {:error, {:insufficient_funds, list(map())}} | {:error, :too_many_inputs}
   @type transaction_t() :: %{
           inputs: nonempty_list(%DB.TxOutput{}),
           outputs: nonempty_list(UtxoSelection.payment_t()),
@@ -113,7 +111,7 @@ defmodule OMG.WatcherInfo.API.Transaction do
     case(
       order.owner
       |> DB.TxOutput.get_sorted_grouped_utxos()
-      |> select_inputs(order)
+      |> TransactionCreator.select_inputs(order)
     ) do
       {:ok, inputs} ->
         TransactionCreator.create(inputs, order)
@@ -121,70 +119,5 @@ defmodule OMG.WatcherInfo.API.Transaction do
       err ->
         err
     end
-  end
-
-  @spec include_typed_data(UtxoSelection.advice_t()) :: UtxoSelection.advice_t()
-  def include_typed_data({:error, _} = err), do: err
-
-  def include_typed_data({:ok, txs}),
-    do: {
-      :ok,
-      %{transactions: Enum.map(txs, fn tx -> Map.put_new(tx, :typed_data, add_type_specs(tx)) end)}
-    }
-
-  # Given an `order`, finds spender's inputs sufficient to perform a payment.
-  # If also provided with receiver's address, creates and encodes a transaction.
-  @spec select_inputs(utxos_map_t(), order_t()) :: inputs_t()
-  defp select_inputs(utxos, %{payments: payments, fee: fee}) do
-    token_utxo_selection =
-      payments
-      |> UtxoSelection.needed_funds(fee)
-      |> UtxoSelection.select_utxo(utxos)
-
-    case UtxoSelection.funds_sufficient(token_utxo_selection) do
-      {:ok, funds} ->
-        stealth_merge_utxos =
-          utxos
-          |> UtxoSelection.prioritize_merge_utxos(funds)
-          |> UtxoSelection.add_utxos_for_stealth_merge(Map.new(funds))
-
-        {:ok, stealth_merge_utxos}
-
-      err ->
-        err
-    end
-  end
-
-  defp add_type_specs(%{inputs: inputs, outputs: outputs, metadata: metadata}) do
-    message =
-      [
-        create_inputs(inputs),
-        create_outputs(outputs),
-        [metadata: metadata || @empty_metadata]
-      ]
-      |> Enum.concat()
-      |> Map.new()
-
-    %{
-      domain: TypedDataHash.Config.domain_data_from_config(),
-      message: message
-    }
-    |> Map.merge(TypedDataHash.Types.eip712_types_specification())
-  end
-
-  defp create_inputs(inputs) do
-    inputs
-    |> Stream.map(fn input -> %{blknum: input.blknum, txindex: input.txindex, oindex: input.oindex} end)
-    |> Stream.concat(Stream.repeatedly(fn -> %{blknum: 0, txindex: 0, oindex: 0} end))
-    |> (fn input -> Enum.zip([:input0, :input1, :input2, :input3], input) end).()
-  end
-
-  defp create_outputs(outputs) do
-    zero_addr = OMG.Eth.zero_address()
-    empty_gen = fn -> %{owner: zero_addr, currency: zero_addr, amount: 0} end
-
-    outputs
-    |> Stream.concat(Stream.repeatedly(empty_gen))
-    |> (fn output -> Enum.zip([:output0, :output1, :output2, :output3], output) end).()
   end
 end
